avoid possible unintended integer overflow in fuzzer.

+++ src/org/sosy_lab/java_smt/test/IntegerTheoryFuzzer.java
@@ -89,7 +89,7 @@ class IntegerTheoryFuzzer {
   }
 
   private IntegerFormula getConstant() {
-    return ifmgr.makeNumber(r.nextInt(2 * maxConstant) - maxConstant);
+    return ifmgr.makeNumber((long) r.nextInt(2 * maxConstant) - maxConstant);
   }
 
   private IntegerFormula getVar() {
